feat: Implement core financial services and calculators for reconciliation, portfolio, budget, and income statements.
