+++ src/services/incomeStatementService.test.ts
@@ -294,53 +294,4 @@ describe('incomeStatementService', () => {
       expect(statement.quarter).toBeUndefined();
     });
   });
-
-  describe('getIncomeCategories', () => {
-    it('should return income categories only', async () => {
-      vi.mocked(transactionService.getTransactionsByPeriod).mockResolvedValue(mockTransactions);
-      vi.mocked(projectService.getProjects).mockResolvedValue(mockProjects);
-
-      const categories = await incomeStatementService.getIncomeCategories(
-        householdId,
-        startDate,
-        endDate,
-      );
-
-      expect(categories.length).toBe(2);
-      expect(categories.every(c => c.items.every(i => i.category.includes('收入')))).toBe(true);
-    });
-  });
-
-  describe('getExpenseCategories', () => {
-    it('should return expense categories only', async () => {
-      vi.mocked(transactionService.getTransactionsByPeriod).mockResolvedValue(mockTransactions);
-      vi.mocked(projectService.getProjects).mockResolvedValue(mockProjects);
-
-      const categories = await incomeStatementService.getExpenseCategories(
-        householdId,
-        startDate,
-        endDate,
-      );
-
-      expect(categories.length).toBe(2);
-      expect(categories.every(c => c.items.every(i => i.category.includes('費用')))).toBe(true);
-    });
-  });
-
-  describe('calculateNetIncome', () => {
-    it('should calculate positive net income', () => {
-      const netIncome = incomeStatementService.calculateNetIncome(120000, 55000);
-      expect(netIncome).toBe(65000);
-    });
-
-    it('should calculate negative net income (loss)', () => {
-      const netIncome = incomeStatementService.calculateNetIncome(50000, 80000);
-      expect(netIncome).toBe(-30000);
-    });
-
-    it('should handle zero values', () => {
-      const netIncome = incomeStatementService.calculateNetIncome(0, 0);
-      expect(netIncome).toBe(0);
-    });
-  });
 });
