+++ src/services/portfolioService.ts
@@ -6,7 +6,6 @@ import {
   query,
   where,
   orderBy,
-  serverTimestamp,
 } from 'firebase/firestore';
 import { db } from '../firebase';
 import {
@@ -15,10 +14,11 @@ import {
   parseWithSchema,
   type Portfolio,
   type PortfolioSnapshot,
-  type PortfolioAccountSnapshot,
+  type AccountSnapshot,
 } from '../schemas';
 import { BaseService } from './baseService';
 import { accountService } from './accountService';
+import { calculatePortfolioSnapshot } from '../domains/finance/calculators/portfolioCalculator';
 
 class PortfolioService extends BaseService<Portfolio> {
   constructor() {
@@ -72,53 +72,25 @@ class PortfolioService extends BaseService<Portfolio> {
     }
 
     // 1. Fetch latest snapshots for linked accounts for the specified year/month
-    const accountSnapshots: PortfolioAccountSnapshot[] = [];
-    let totalValue = 0;
+    const accountSnapshots = new Map<string, AccountSnapshot | null>();
+    const accounts = [];
 
     for (const accountId of portfolio.accountIds) {
       const account = await accountService.getAccount(householdId, accountId);
       if (!account) continue;
+      
+      accounts.push(account);
 
       const snapshots = await accountService.getSnapshots(householdId, accountId, year, month);
-      // Use the snapshot for this specific month if available, otherwise...
-      // Actually, for a portfolio snapshot of Jan 2025, we should strictly look for Jan 2025 account snapshots.
-      // If missing, maybe we shouldn't include it or throw error?
-      // For now, let's assume if it's missing, value is 0 or we skip.
-      // User requirement: "先記錄各 Account 的 Snapshot, 再建立 Portfolio 的 Snapshot"
-      // So we expect them to exist.
-
+      
       if (snapshots.length > 0) {
-        const snapshot = snapshots[0]; // getSnapshots returns sorted desc, so first is latest for that month
-        const accountSnapshot: PortfolioAccountSnapshot = {
-          accountId: account.id,
-          accountName: account.name,
-          type: account.type,
-          value: snapshot.amount,
-        };
-        
-        // Only include holdings if they exist
-        if (snapshot.holdings && snapshot.holdings.length > 0) {
-          accountSnapshot.holdings = snapshot.holdings;
-        }
-        
-        accountSnapshots.push(accountSnapshot);
-        totalValue += snapshot.amount;
+        accountSnapshots.set(accountId, snapshots[0]);
       } else {
-        // If no snapshot for this month, maybe try to get latest?
-        // But that might be misleading for "Performance of Jan 2025".
-        // Let's just include the account with 0 value or skip?
-        // Better to include with 0 to show it's part of portfolio but missing data.
-        accountSnapshots.push({
-          accountId: account.id,
-          accountName: account.name,
-          type: account.type,
-          value: 0,
-        });
+        accountSnapshots.set(accountId, null);
       }
     }
 
-    // 2. Calculate Performance
-    // Need previous month's portfolio snapshot
+    // 2. Fetch previous month's portfolio snapshot
     let prevYear = year;
     let prevMonth = month - 1;
     if (prevMonth === 0) {
@@ -129,46 +101,17 @@ class PortfolioService extends BaseService<Portfolio> {
     const prevSnapshots = await this.getSnapshots(householdId, portfolioId, prevYear, prevMonth);
     const prevSnapshot = prevSnapshots.length > 0 ? prevSnapshots[0] : null;
 
-    const openingValue = prevSnapshot ? prevSnapshot.performance.closingValue : 0;
-    const closingValue = totalValue;
-    const netCashFlow = cashFlow.deposits - cashFlow.withdrawals;
-    
-    // Gain = Closing - Opening - NetFlow
-    const gain = closingValue - openingValue - netCashFlow;
-
-    // Return Rate
-    // User formula: gain / (openingValue + deposits) * 100
-    let returnRate = 0;
-    const denominator = openingValue + cashFlow.deposits;
-    if (denominator > 0) {
-      returnRate = (gain / denominator) * 100;
-    }
-
-    // Cumulative Gain
-    const prevCumulativeGain = prevSnapshot ? prevSnapshot.performance.cumulativeGain : 0;
-    const cumulativeGain = prevCumulativeGain + gain;
-    
-    // Cumulative Return Rate
-    // User formula: (cumulativeGain / totalInvested) * 100
-    // We need totalInvested.
-    // totalInvested = currentTotalValue - cumulativeGain
-    // Because: cumulativeGain = totalValue - totalInvested
-    const totalInvested = closingValue - cumulativeGain;
-    
-    let cumulativeReturnRate = 0;
-    if (totalInvested > 0) {
-      cumulativeReturnRate = (cumulativeGain / totalInvested) * 100;
-    }
-
-    const performance = {
-      openingValue,
-      closingValue,
-      netCashFlow,
-      gain,
-      returnRate,
-      cumulativeGain,
-      cumulativeReturnRate,
-    };
+    // 3. Calculate Snapshot
+    const snapshotData = calculatePortfolioSnapshot({
+      year,
+      month,
+      portfolioId,
+      accounts,
+      accountSnapshots,
+      prevSnapshot,
+      cashFlow,
+      createdBy,
+    });
 
     const snapshotRef = doc(
       collection(db, 'households', householdId, 'portfolios', portfolioId, 'snapshots'),
@@ -176,20 +119,11 @@ class PortfolioService extends BaseService<Portfolio> {
     const snapshotId = snapshotRef.id;
 
     const newSnapshot = {
+      ...snapshotData,
       id: snapshotId,
-      year,
-      month,
-      accounts: accountSnapshots,
-      totalValue,
-      cashFlow,
-      performance,
-      createdBy,
-      createdAt: serverTimestamp(),
     };
-    console.log('newSnapshot', newSnapshot);
 
     await setDoc(snapshotRef, newSnapshot);
-    console.log('snapshotId', snapshotId);
     return snapshotId;
   }
